ner_set 추가
